Refactor tests to use new built-in method for reading logs
