+++ test/integration/specs/api-abort-on-error-local/time-out.specs.js
@@ -1,15 +1,11 @@
 const test = require('../../../../index')
-const utils = require('../utils')
+const utils = require('../../../../utils')
 
 test.describe('api-abort-on-error local tests execution', () => {
   let outerrLog
 
-  test.before((done) => {
-    utils.readOutErr()
-      .then((log) => {
-        outerrLog = log
-        done()
-      })
+  test.before(async () => {
+    outerrLog = await utils.logs.combined('package-test')
   })
 
   test.it('should have exited api with an error', () => {
