+++ test/integration/specs/exit/exit-ko.specs.js
@@ -1,16 +1,12 @@
 
 const test = require('../../../../index')
-const utils = require('../utils')
+const utils = require('../../../../utils')
 
 test.describe('package test execution', () => {
   let exitCodeLog
 
-  test.before((done) => {
-    utils.readExitCode()
-      .then((log) => {
-        exitCodeLog = log
-        done()
-      })
+  test.before(async () => {
+    exitCodeLog = await utils.logs.exitCode('package-test')
   })
 
   test.it('should finish with an exit code different to 0', () => {
