+++ test/integration/specs/log-levels/api-log.specs.js
@@ -2,18 +2,14 @@ const _ = require('lodash')
 
 const test = require('../../../../index')
 
-const utils = require('../utils')
+const utils = require('../../../../utils')
 const definitions = require('./logs-definitions')
 
 test.describe('services logs', () => {
   let outerrLog
 
-  test.before((done) => {
-    utils.readOutErr()
-      .then((log) => {
-        outerrLog = log
-        done()
-      })
+  test.before(async () => {
+    outerrLog = await utils.logs.combined('package-test')
   })
 
   const expects = (logDefinitions, log) => {
